telegram-chatbot/requirements.txt
python-telegram-bot
google-generativeai
python-dotenv
mcp
httpx
